+++ src/model/model_molecule.hpp
@@ -55,10 +55,10 @@ namespace VTX
 				_bonds.insert( _bonds.end(), p_bonds.begin(), p_bonds.end() );
 			}
 
-			inline uint getChainCount() const { return _chains.size(); }
-			inline uint getResidueCount() const { return _residues.size(); }
-			inline uint getAtomCount() const { return _atoms.size(); }
-			inline uint getBondCount() const { return _bonds.size(); }
+			inline uint getChainCount() const { return (uint)_chains.size(); }
+			inline uint getResidueCount() const { return (uint)_residues.size(); }
+			inline uint getAtomCount() const { return (uint)_atoms.size(); }
+			inline uint getBondCount() const { return (uint)_bonds.size(); }
 
 			inline Math::AABB & AABB() { return _aabb; }
